# apps/api/internal/docs/swagger.yaml
definitions:
  models.User:
    properties:
      address:
        type: string
      age:
        type: integer
      email:
        example: sitthiporn.po@gmail.com
        type: string
      id:
        description: gorm.Model
        example: 8
        type: integer
      name:
        type: string
      passwordHash:
        type: string
      passwordSalt:
        type: string
      surname:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
paths:
  /:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Welcome to Fiber API
      tags:
      - Root
  /api/users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
      summary: Get All Users
      tags:
      - Users
  /api/users/email/{email}:
    get:
      consumes:
      - application/json
      description: Get a user by their email address
      parameters:
      - description: The email of the resource Sitthiporn.po@gmail.com
        in: path
        name: Sitthiporn.po@gmail.coml
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
      summary: Get User by Email
      tags:
      - Users
  /api/users/id/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: The ID of the user
        in: path
        name: string
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
      summary: Get All Users
      tags:
      - Users
swagger: "2.0"